+++ aesara/compile/ops.py
@@ -11,8 +11,9 @@ import warnings
 from typing import Dict, Tuple
 
 from aesara.graph.basic import Apply
-from aesara.graph.op import COp, Op
+from aesara.graph.op import Op
 from aesara.graph.type import CType
+from aesara.link.c.op import COp
 
 
 def register_view_op_c_code(type, code, version=()):
