+++ aesara/gpuarray/blas.py
@@ -10,9 +10,9 @@ from aesara.gpuarray.basic_ops import (
 )
 from aesara.gpuarray.opt_util import inplace_allocempty
 from aesara.graph.basic import Apply
-from aesara.graph.op import _NoPythonCOp
 from aesara.graph.opt import LocalOptGroup, in2out
 from aesara.graph.params_type import ParamsType
+from aesara.link.c.op import _NoPythonCOp
 from aesara.scalar import bool as bool_t
 from aesara.tensor.basic import as_tensor_variable
