+++ aesara/tensor/blas_c.py
@@ -1,7 +1,7 @@
 from aesara.configdefaults import config
-from aesara.graph.op import COp
 from aesara.graph.opt import in2out
 from aesara.graph.params_type import ParamsType
+from aesara.link.c.op import COp
 from aesara.scalar import bool as bool_t
 from aesara.tensor import basic as at
 from aesara.tensor.blas import (
