+++ aesara/tensor/nnet/ctc.py
@@ -5,9 +5,9 @@ import aesara.tensor as at
 from aesara.configdefaults import config
 from aesara.gradient import grad_undefined
 from aesara.graph.basic import Apply
-from aesara.graph.op import ExternalCOp, OpenMPOp
 from aesara.graph.opt import local_optimizer
 from aesara.link.c.cmodule import GCC_compiler
+from aesara.link.c.op import ExternalCOp, OpenMPOp
 from aesara.tensor.basic_opt import register_canonicalize
 from aesara.tensor.blas import batched_dot
 from aesara.tensor.extra_ops import cpu_contiguous
