+++ setup.cfg
@@ -151,6 +151,10 @@ check_untyped_defs = False
 ignore_errors = True
 check_untyped_defs = False
 
+[mypy-aesara.link.c.op]
+ignore_errors = True
+check_untyped_defs = False
+
 [mypy-aesara.link.utils]
 ignore_errors = True
 check_untyped_defs = False
