+++ tests/compile/test_debugmode.py
@@ -17,9 +17,10 @@ from aesara.compile.mode import predefined_modes
 from aesara.configdefaults import config
 from aesara.graph.basic import Apply, Variable
 from aesara.graph.features import BadOptimization
-from aesara.graph.op import COp, Op
+from aesara.graph.op import Op
 from aesara.graph.opt import local_optimizer
 from aesara.graph.optdb import EquilibriumDB
+from aesara.link.c.op import COp
 from aesara.tensor.math import add, dot, log
 from aesara.tensor.type import TensorType, dvector, fmatrix, fvector, vector
 from tests import unittest_tools as utt
