+++ tests/graph/test_compute_test_value.py
@@ -9,8 +9,9 @@ from aesara import scalar as aes
 from aesara.configdefaults import config
 from aesara.graph import utils
 from aesara.graph.basic import Apply
-from aesara.graph.op import COp, Op
+from aesara.graph.op import Op
 from aesara.graph.type import Type
+from aesara.link.c.op import COp
 from aesara.tensor.math import _allclose, dot
 from aesara.tensor.type import fmatrix, iscalar, matrix, vector
